Move feedforward out of tensors
